requirements.txt
# 核心依赖
litellm>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# 向量存储
faiss-cpu>=1.7.4
numpy>=1.24.0

# 代码解析
tree-sitter>=0.21.0
tree-sitter-java>=0.21.0

# 工具
tqdm>=4.65.0
pyyaml>=6.0
